challenge component 수정
